Update pipeline to run on Docker runners
The runners connected to this GitLab instance are now Docker runners
instead of shell runners.  This means an image should be specified to
run the commands in.

States will run using the kitchen-base image.  This is a Ruby image with
the Docker binary available to communicate with the Docker daemon on
the host.