Small refactor
Move the image declaration to the top of the file, outside of the test
stanza to avoid needing to redeclare the image for every job.

Also removed the stages declaration and generalised before_script to all
jobs as the defaults of build, test and deploy are suitable for this
project and the before_script commands will be useful for any future
jobs which may be added.